preload the arcgis api for javascript on the index route
preload JSAPI after the application template is rendered

remove map controller and empty the index route

add preload example to README
