+++ tests/dummy/app/styles/app.css
@@ -6,7 +6,7 @@
 
 /* app styles */
 body {
-  padding-top: 20px;
+  padding-top: 10px;
   padding-bottom: 20px;
 }
