+++ tests/dummy/app/templates/application.hbs
@@ -1,5 +1,9 @@
 <div class="container">
 
+  <!-- JSAPI load status -->
+  <p style="text-align: right">The ArcGIS API for JavaScript
+    is: <strong>{{#unless jsapiLoaded }}<span class="text-muted">Loading...</span>{{else}}<span class="text-success">Loaded</span>{{/unless}}</strong></p>
+
   <!-- navbar -->
   <nav class="navbar navbar-default">
     <div class="container-fluid">
