+++ tests/dummy/app/templates/index.hbs
@@ -1,8 +1,6 @@
 <h2>ember-esri-loader</h2>
-<p style="margin-top: 1em;">The ArcGIS API for JavaScript
-  is: <strong>{{unless jsapiLoaded "NOT"}} loaded</strong>.</p>
 <p>If you navigated directly to this route, then the <a href="https://developers.arcgis.com/javascript/" target="_blank">ArcGIS API for JavaScript</a>
-  will not be loaded on the page until you navigate to the {{#link-to 'map'}}map
+  will not be loaded on the page until after the template has rendered or until you navigate to the {{#link-to 'map'}}map
   route{{/link-to}}. This helps keep the initial load time down on non-map routes.</p>
 <p>
   If you transitioned here from the {{#link-to 'map'}}map{{/link-to}} route, then the ArcGIS API for JavaScript will already be loaded.
