+++ tests/dummy/app/templates/map.hbs
@@ -1,4 +1,2 @@
 <h2>Map</h2>
-<p style="margin-top: 1em;">The ArcGIS API for JavaScript
-  is: <strong>{{unless jsapiLoaded "NOT"}} loaded</strong>.</p>
 {{esri-map}}
